+++ plugin.php
@@ -4,7 +4,7 @@
  *
  * @wp-plugin
  *
- * Version: 160828.39509
+ * Version: 160828.49551
  * Text Domain: woocommerce-kb-articles
  * Plugin Name: WooCommerce KB Articles Pro
  *